% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.fit.2.R
\docType{data}
\name{def.fit.2}
\alias{def.fit.2}
\title{Default fit of order 2 for medals}
\format{An object of class \code{glm} (inherits from \code{lm}) of length 21.}
\usage{
def.fit.2
}
\description{
Default model fit from training data for MEDALS package. This fit used 9 subjects with 4 imaging modalities (FLAIR, T1w,T2w,DWI), and carried out a 2nd order MEDALS analysis using the first 8 PCs as predictors.
}
\keyword{datasets}
