% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.fit.R
\docType{data}
\name{def.fit}
\alias{def.fit}
\title{Default fit for medals}
\format{An object of class \code{glm} (inherits from \code{lm}) of length 21.}
\usage{
def.fit
}
\description{
Testing
}
\keyword{datasets}

