% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.loads.2.R
\docType{data}
\name{def.loads.2}
\alias{def.loads.2}
\title{Default loadings for 2nd order medals}
\format{An object of class \code{matrix} with 216 rows and 216 columns.}
\usage{
def.loads.2
}
\description{
Default loadings matrix from training data for MEDALS package. This fit used 9 subjects with 4 imaging modalities (FLAIR, T1w,T2w,DWI), and carried out a 2nd order MEDALS analysis
}
\keyword{datasets}
