% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.img.moment.dat.R
\name{get.img.moment.dat}
\alias{get.img.moment.dat}
\title{A function to calculate the matrix of neighborhood moments for a subject MEDALS}
\usage{
get.img.moment.dat(imgs.path, mask.path, mpower = 4, verbose = TRUE)
}
\arguments{
\item{imgs.path}{A list of paths to images for one subject.}

\item{mask.path}{A path to the braiin mask for each subject.}

\item{mpower}{A scalar specifying the highest moment wanted for the MEDALS analysis.}

\item{verbose}{print diagnostic messages}
}
\description{
This function allows you to calculate the matrix of neighborhood moments for MEDALS.
}
\keyword{MEDALS,}
\keyword{Segmentation}
\keyword{Sufficiency,}

