% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.loadings.R
\name{get.loadings}
\alias{get.loadings}
\title{A function to create the loadings matrix for  MEDALS.}
\usage{
get.loadings(cov.mat)
}
\arguments{
\item{cov.mat}{The crossproduct matrix ($X^TX$) for decomposition to get loadings. Should be the output of imaging.cp.mat()}
}
\description{
This function allows you to create loadings matrix for PC scores in the MEDALS pipeline. T
}

