% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imaging.cp.mat.R
\name{imaging.cp.mat}
\alias{imaging.cp.mat}
\title{A function to create cross-product matrix for MEDALS}
\usage{
imaging.cp.mat(path.img.list, path.mask.list, mean.vec, sd.vec)
}
\arguments{
\item{path.img.list}{A list of sublists, where each sublist is a list of paths to the files for each imaging modality (ex. T1w, T2w, DWI, etc.) for a particular subject. This has to contain at least one sublist}

\item{path.mask.list}{A list of paths to the brain mask for each subject. This should be in the same order as the sublists for path.img.list}

\item{mean.vec}{A vector of column means of $X$. This will be used to center columns before the PCA.}

\item{sd.vec}{A vector of column standard deviations of $X$. This will be used to scale columns before the PCA.}
}
\description{
This function allows you to create the cross-product matrix ($X^TX$) for MEDALS. This is the matrix the PCA will be performed on. This is the 2nd step in the MEDALS pipeline.
}
\examples{
imaging.cp()
}
\keyword{MEDALS,}
\keyword{PCA,}
\keyword{Segmentation}

