% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment.cor.mat.R
\name{moment.cor.mat}
\alias{moment.cor.mat}
\title{A function to calculate sufficient statistics and out cross-product matrix efficiently for medals}
\usage{
moment.cor.mat(path.img.list, path.mask.list, mpower = 4, verbose = TRUE)
}
\arguments{
\item{path.img.list}{A list of sublists, where each sublist is a list of paths to the files for each imaging modality (ex. T1w, T2w, DWI, etc.) for a particular subject. This has to contain at least one sublist}

\item{path.mask.list}{A list of paths to the brain mask for each subject. This should be in the same order as the sublists for path.img.list}

\item{mpower}{A scalar specifying the highest moment wanted for the MEDALS analysis. (Should be at least 1)}

\item{verbose}{print diagnostic messages}
}
\description{
This function allows you to calculate the sufficient statistics to center and scale each column in the huge (long) matrix X for medals, then outputs the center and scaled cross product matrices. This should be the first step of the MEDALS pipeline
}
\keyword{MEDALS,}
\keyword{Segmentation}
\keyword{Sufficiency,}
