% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc.var.R
\name{pc.var}
\alias{pc.var}
\title{A function to calculate variance explained by each Principal Component for MEDALS}
\usage{
pc.var(mat, n)
}
\arguments{
\item{mat}{The crossproduct matrix, this can be the output of imaging.cp}

\item{n}{The number of observations in the original matrix. This can be the output for imaging.suff.stat}
}
\value{
List of 2 elements: \code{var}: variance of each
component, \code{cum_pct}: cumulative percent variance explained
}
\description{
This function returns the variance explained by each PC and the cumulative variance explained in the medals package, this should be used with the output of imaging.cp
}
\keyword{MEDALS,}
\keyword{variance}

