% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.fit.R
\docType{data}
\name{def.fit}
\alias{def.fit}
\title{Default fit for medals}
\format{An object of class \code{glm} (inherits from \code{lm}) of length 21.}
\usage{
def.fit
}
\description{
Default model fit from training data for MEDALS package. This fit used 9 subjects with 4 imaging modalities (FLAIR, T1w,T2w,DWI), and carried out a 4th order MEDALS analysis using the first 8 PCs as predictors.
}
\keyword{datasets}
