% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.loads.R
\docType{data}
\name{def.loads}
\alias{def.loads}
\title{Default loadings for medals}
\format{An object of class \code{matrix} with 432 rows and 432 columns.}
\usage{
def.loads
}
\description{
Default loadings matrix from training data for MEDALS package. This fit used 9 subjects with 4 imaging modalities (FLAIR, T1w,T2w,DWI), and carried out a 4th order MEDALS analysis
}
\keyword{datasets}
