% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imaging.cp.mat.R
\name{imaging.cp.mat}
\alias{imaging.cp.mat}
\title{A function to create cross-product matrix for MEDALS}
\usage{
imaging.cp.mat(mean.vec, sd.vec, n.vec, cp.list)
}
\arguments{
\item{mean.vec}{A vector of column means of $X$. This will be used to center columns before the PCA.}

\item{sd.vec}{A vector of column standard deviations of $X$. This will be used to scale columns before the PCA.}

\item{n.vec}{A vector containing the number of voxels in the brain mask for subject}

\item{cp.list}{A list of subject-specific cross-product matrices}
}
\description{
This function allows you to create the cross-product matrix ($X^TX$) for MEDALS. This is the matrix the PCA will be performed on. This is the 2nd step in the MEDALS pipeline.
}
\keyword{MEDALS,}
\keyword{PCA,}
\keyword{Segmentation}

