% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.pred.img.R
\name{make.pred.img}
\alias{make.pred.img}
\title{A function to make prediction images for MEDALS}
\usage{
make.pred.img(score.img.list, path.mask.list, fit = medals::def.fit,
  verbose = TRUE)
}
\arguments{
\item{score.img.list}{A list of sublists, where each sublist is a list of images created from the PC scores. This can be the output of make.score.img().This has to contain at least one sublist}

\item{path.mask.list}{A list of paths to the brain mask for each subject. This should be in the same order as the sublists for score.img.list}

\item{fit}{A model fit object, either use the default or use the output from get.model.fit().}

\item{verbose}{Print diagnostic messages}
}
\description{
This function allows you to create prediction images from the MEDALS pipeline. This should be the last step
}
\keyword{MEDALS,}
\keyword{Prediction}
